requirements.txt
linear==0.0.dev0
matplotlib==3.4.3
model==0.6.0
numpy==1.22.3
opencv_python==4.5.5.62
opencv_python_headless==4.5.5.64
pandas==1.3.3
Pillow==9.5.0
PyYAML==6.0
scikit_learn==0.24.2
scipy==1.7.1
tensorboardX==2.6
tifffile==2021.11.2
timm==0.9.2
torch==1.12.0
torchvision==0.13.0
visualize==0.5.0
visualizer==0.0.10
